@@ -18,9 +18,13 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
+  dio: ^4.0.5
+  retrofit: '>=3.0.0 <4.0.0'
+  logger: any 
+  pretty_dio_logger: ^1.1.1
   flutter:
     sdk: flutter
 
@@ -32,6 +36,10 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  retrofit_generator: ^4.0.1
+  build_runner: ^2.1.8
+  json_annotation: ^4.4.0
+  json_serializable: ^6.1.5
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
